+++ kore/src/Kore/Step/Axiom/EvaluationStrategy.hs
@@ -17,35 +17,21 @@ module Kore.Step.Axiom.EvaluationStrategy
 
 import qualified Control.Monad as Monad
 import qualified Data.Foldable as Foldable
-import           Data.Function
 import           Data.Maybe
                  ( isJust )
 import qualified Data.Text as Text
 import qualified Data.Text.Prettyprint.Doc as Pretty
 
-import qualified Kore.Attribute.Axiom as Attribute.Axiom
-import qualified Kore.Attribute.Axiom.Concrete as Axiom.Concrete
 import qualified Kore.Attribute.Symbol as Attribute
-import           Kore.Internal.Conditional
-                 ( andCondition )
 import qualified Kore.Internal.MultiOr as MultiOr
                  ( extractPatterns )
 import qualified Kore.Internal.OrPattern as OrPattern
-import           Kore.Internal.Pattern
-                 ( Pattern )
-import qualified Kore.Internal.Pattern as Pattern
 import           Kore.Internal.Symbol
 import           Kore.Internal.TermLike
-import qualified Kore.Internal.TermLike as TermLike
 import qualified Kore.Proof.Value as Value
-import           Kore.Step.Axiom.Matcher
-                 ( matchIncremental )
-import           Kore.Step.Remainder
-                 ( ceilChildOfApplicationOrTop )
-import qualified Kore.Step.Result as Result
+import           Kore.Step.Axiom.Evaluate
 import           Kore.Step.Rule
-                 ( EqualityRule (EqualityRule, getEqualityRule) )
-import qualified Kore.Step.Rule as RulePattern
+                 ( EqualityRule )
 import           Kore.Step.Simplification.Data
                  ( AttemptedAxiom,
                  AttemptedAxiomResults (AttemptedAxiomResults),
@@ -54,13 +40,7 @@ import           Kore.Step.Simplification.Data
 import qualified Kore.Step.Simplification.Data as AttemptedAxiomResults
                  ( AttemptedAxiomResults (..) )
 import qualified Kore.Step.Simplification.Data as AttemptedAxiom
-                 ( AttemptedAxiom (..), hasRemainders, maybeNotApplicable )
-import qualified Kore.Step.Simplification.OrPattern as OrPattern
-                 ( simplifyPredicatesWithSmt )
-import           Kore.Step.Step
-                 ( UnificationProcedure (UnificationProcedure) )
-import qualified Kore.Step.Step as Step
-import qualified Kore.Unification.Unify as Monad.Unify
+                 ( AttemptedAxiom (..), hasRemainders )
 import           Kore.Unparser
                  ( Unparse, unparse )
 import           Kore.Variables.Fresh
@@ -83,9 +63,7 @@ that define it.
 definitionEvaluation
     :: [EqualityRule Variable]
     -> BuiltinAndAxiomSimplifier
-definitionEvaluation rules =
-    BuiltinAndAxiomSimplifier
-        (evaluateWithDefinitionAxioms rules)
+definitionEvaluation rules = BuiltinAndAxiomSimplifier (evaluateAxioms rules)
 
 {- | Creates an evaluator for a function from all the rules that define it.
 
@@ -127,7 +105,7 @@ totalDefinitionEvaluation rules =
             else return result
       where
         evaluate =
-            evaluateWithDefinitionAxioms
+            evaluateAxioms
                 rules
                 predicateSimplifier
                 termSimplifier
@@ -290,86 +268,3 @@ applyFirstSimplifierThatWorks
                 simplifier
                 axiomIdToSimplifier
                 patt
-
-evaluateWithDefinitionAxioms
-    :: forall variable simplifier
-    .   ( FreshVariable variable
-        , SortedVariable variable
-        , Show variable
-        , Unparse variable
-        , MonadSimplify simplifier
-        )
-    => [EqualityRule Variable]
-    -> PredicateSimplifier
-    -> TermLikeSimplifier
-    -> BuiltinAndAxiomSimplifierMap
-    -- ^ Map from axiom IDs to axiom evaluators
-    -> TermLike variable
-    -> simplifier (AttemptedAxiom variable)
-evaluateWithDefinitionAxioms
-    definitionRules
-    _predicateSimplifier
-    _termSimplifier
-    _axiomSimplifiers
-    patt
-  | any ruleIsConcrete definitionRules
-  , not (TermLike.isConcrete patt)
-  = return AttemptedAxiom.NotApplicable
-  | otherwise
-  = AttemptedAxiom.maybeNotApplicable $ do
-    let
-        -- TODO (thomas.tuegel): Figure out how to get the initial conditions
-        -- and apply them here, to remove remainder branches sooner.
-        expanded :: Pattern variable
-        expanded = Pattern.fromTermLike patt
-
-    results <- applyRules expanded (map unwrapEqualityRule definitionRules)
-    Monad.guard (any Result.hasResults results)
-    mapM_ rejectNarrowing results
-
-    ceilChild <- ceilChildOfApplicationOrTop patt
-    let
-        result =
-            Result.mergeResults results
-            & Result.mapConfigs
-                keepResultUnchanged
-                ( markRemainderEvaluated
-                . introduceDefinedness ceilChild
-                )
-        keepResultUnchanged = id
-        introduceDefinedness = flip andCondition
-        markRemainderEvaluated = fmap mkEvaluated
-
-    AttemptedAxiomResults
-        <$> OrPattern.simplifyPredicatesWithSmt (Step.gatherResults result)
-        <*> OrPattern.simplifyPredicatesWithSmt (Step.remainders result)
-
-  where
-    ruleIsConcrete =
-        Axiom.Concrete.isConcrete
-        . Attribute.Axiom.concrete
-        . RulePattern.attributes
-        . getEqualityRule
-
-    unwrapEqualityRule (EqualityRule rule) =
-        RulePattern.mapVariables fromVariable rule
-
-    rejectNarrowing (Result.results -> results) =
-        (Monad.guard . not) (Foldable.any Step.isNarrowingResult results)
-
-    applyRules initial rules =
-        Monad.Unify.maybeUnifierT
-        $ Step.applyRulesSequence unificationProcedure initial rules
-
-    ignoreUnificationErrors unification pattern1 pattern2 =
-        Monad.Unify.runUnifierT (unification pattern1 pattern2)
-        >>= either (couldNotMatch pattern1 pattern2) Monad.Unify.scatter
-
-    couldNotMatch pattern1 pattern2 _ =
-        Monad.Unify.explainAndReturnBottom
-            "Could not match patterns"
-            pattern1
-            pattern2
-
-    unificationProcedure =
-        UnificationProcedure (ignoreUnificationErrors matchIncremental)
